Share fixtures across module
